Remove SourceStore
